test: move parent-death regression to unix file
