+++ internal/runtime/builtin/command/command_test.go
@@ -10,14 +10,11 @@ import (
 	"os/exec"
 	"path/filepath"
 	goruntime "runtime"
-	"strconv"
 	"strings"
 	"sync"
-	"syscall"
 	"testing"
 	"time"
 
-	"github.com/dagucloud/dagu/internal/cmn/cmdutil"
 	"github.com/dagucloud/dagu/internal/core"
 	"github.com/dagucloud/dagu/internal/core/spec"
 	"github.com/dagucloud/dagu/internal/runtime"
@@ -395,100 +392,6 @@ func TestCommandExecutor_Kill(t *testing.T) {
 	executor = &commandExecutor{cmd: &exec.Cmd{}}
 	err = executor.Kill(os.Interrupt)
 	assert.NoError(t, err)
-}
-
-func TestCommandExecutor_CleansProcessGroupWhenParentDies(t *testing.T) {
-	if goruntime.GOOS == "windows" {
-		t.Skip("Skipping Unix-specific test on Windows")
-	}
-	if os.Getenv("DAGU_COMMAND_PARENT_DEATH_HELPER") == "1" {
-		runCommandParentDeathHelper(t)
-		return
-	}
-
-	tmpDir := t.TempDir()
-	pidFile := filepath.Join(tmpDir, "script.pid")
-	readyFile := filepath.Join(tmpDir, "helper.ready")
-
-	cmd := exec.Command(os.Args[0], "-test.run=TestCommandExecutor_CleansProcessGroupWhenParentDies")
-	cmd.Env = append(os.Environ(),
-		"DAGU_COMMAND_PARENT_DEATH_HELPER=1",
-		"DAGU_COMMAND_PARENT_DEATH_DIR="+tmpDir,
-	)
-	cmdutil.SetupCommand(cmd)
-
-	require.NoError(t, cmd.Start())
-	t.Cleanup(func() {
-		if cmd.Process != nil {
-			_ = cmdutil.KillProcessGroup(cmd, os.Kill)
-		}
-	})
-
-	require.Eventually(t, func() bool {
-		_, err := os.Stat(readyFile)
-		return err == nil
-	}, 5*time.Second, 25*time.Millisecond, "helper did not start command script")
-
-	pidData, err := os.ReadFile(pidFile)
-	require.NoError(t, err)
-	scriptPID, err := strconv.Atoi(strings.TrimSpace(string(pidData)))
-	require.NoError(t, err)
-	t.Cleanup(func() {
-		_ = syscall.Kill(-scriptPID, syscall.SIGKILL)
-	})
-
-	require.NoError(t, cmdutil.KillProcessGroup(cmd, os.Kill))
-	_ = cmd.Wait()
-
-	require.Eventually(t, func() bool {
-		return !processGroupAlive(scriptPID)
-	}, 2*time.Second, 25*time.Millisecond, "script process group should die with its parent")
-}
-
-func runCommandParentDeathHelper(t *testing.T) {
-	t.Helper()
-
-	tmpDir := os.Getenv("DAGU_COMMAND_PARENT_DEATH_DIR")
-	require.NotEmpty(t, tmpDir)
-	pidFile := filepath.Join(tmpDir, "script.pid")
-	readyFile := filepath.Join(tmpDir, "helper.ready")
-
-	step := core.Step{
-		Name: "parent-death",
-		Dir:  tmpDir,
-		Script: "echo $$ > " + pidFile + "\n" +
-			"while true; do sleep 1; done\n",
-	}
-
-	ctx := setupTestContext(t, nil, step)
-	env := runtime.GetEnv(ctx)
-	env.WorkingDir = tmpDir
-	ctx = runtime.WithEnv(ctx, env)
-
-	commandExec, err := NewCommand(ctx, step)
-	require.NoError(t, err)
-	errCh := make(chan error, 1)
-	go func() {
-		errCh <- commandExec.Run(ctx)
-	}()
-
-	require.Eventually(t, func() bool {
-		_, err := os.Stat(pidFile)
-		return err == nil
-	}, 5*time.Second, 25*time.Millisecond, "command script did not start")
-	require.NoError(t, os.WriteFile(readyFile, []byte("ok"), 0600))
-
-	select {
-	case err := <-errCh:
-		require.NoError(t, err)
-	case <-time.After(30 * time.Second):
-		t.Fatal("helper was not killed by parent test")
-	}
-}
-
-func processGroupAlive(pid int) bool {
-	err := syscall.Kill(-pid, 0)
-	return err == nil || err == syscall.EPERM
 }
 
 func TestCommandConfig_NewCmd(t *testing.T) {
